Bug fix of ble hid buffer and USB device control
The index of ble hid report index written to buffer is not assigned when buffer is not empty, will cause error when it dequeue.
Computer will disconnect the USB device which long time do not report anything. If the keyboard USB device activate when the cable plug in while not using the USB protocol to transfer key codes, this will happen and cannot reconnect unless you re-plug the cable. To solve the problem, the USB device will only activate when 1. using USB protocol and 2. connecting to the computer
Added the USB started check when sending key codes via USB protocol