Timer related optimization
Use app_timer as keyboard task drive timer and rewrite the program of indicator led.
if a high freq app_timer is config as repeat mode, it will fill up scheduler and cause NO_MEM_ERROR when there is a long interrupt with higher priority. A better way to deal with it is to use single-shot mode and start the timer at the end of its timeout event handler. So only after the timeout handler function is executed, can a new entry be enqueued into the scheduler. This may loose some accuracy of the interval but who cares ?